re-organised components to BestSection and ProductSection, then merged PhotoCard component into ProductList.
